多轮自动化：超远tp自动多步平移（FarRelocate.Request long化+客户端循环translate，一次重载）；auto_relocate默认true（多轮链：边缘自动重定位）

+++ mod/src/main/java/com/farlands/g1/FarRelocate.java
@@ -10,17 +10,17 @@ package com.farlands.g1;
 public final class FarRelocate {
 
     public static final class Request {
-        public final int dx;
-        public final int dz;
+        public final long dx;
+        public final long dz;
         /** New epoch to persist after the shift (NaN = keep current). */
         public final double newEpochX;
         public final double newEpochZ;
 
-        public Request(int dx, int dz) {
+        public Request(long dx, long dz) {
             this(dx, dz, Double.NaN, Double.NaN);
         }
 
-        public Request(int dx, int dz, double newEpochX, double newEpochZ) {
+        public Request(long dx, long dz, double newEpochX, double newEpochZ) {
             this.dx = dx;
             this.dz = dz;
             this.newEpochX = newEpochX;
